+++ tests/conftest.py
@@ -58,7 +58,9 @@ def check_env_variables():
 def reset_logger():
     def reset():
         loguru.logger.remove()
-        loguru.logger.__init__(loguru._logger.Core(), None, 0, False, False, False, False, None, {})
+        loguru.logger.__init__(
+            loguru._logger.Core(), None, 0, False, False, False, False, True, None, {}
+        )
         loguru._logger.context.set({})
         logging.Logger.manager.loggerDict.clear()
         logging.root = logging.RootLogger("WARNING")